+++ include/hpp/bezier-com-traj/waypoints/waypoints_definition.hh
@@ -113,7 +113,7 @@ bezier_wp_t::t_point_t computeWwaypoints(const ProblemData& pData, double T);
 
 coefs_t computeFinalVelocityPoint(const ProblemData& pData, double T);
 
-size_t dimVar(const ProblemData& pData);
+int dimVar(const ProblemData& pData);
 
 std::pair<MatrixXX, VectorX> computeVelocityCost(const ProblemData& pData, double T,
                                                  std::vector<bezier_t::point_t> pi = std::vector<bezier_t::point_t>());
